Restore app service AFTER there are replicas available
We were doing this at the very beginning. This is odd as if a user
for whatever reason refresh the page before the app pods didn't start
up and are available they'll see a `503 Service Unavailable`.

It makes sense for the app service to point to the unidler host until
the unidling process is completed successfully.

Also, if the unidling fails the service will incorrectly point to
the not-quite-unidled app.