Fixed logs wording to be more accurate
- "deployment DON'T HAVE annotation" when annotation not found
- simplified logic to ignore "nonexistent key" (see note)
- improved logging when "nonexistent" key error happen to be explicit about
  the fact we're ignoring this error
- Only show "Successfully Patched" log message when Patch genuenly returned
  a `200 OK`

Note on "nonexistent key"
=========================
the "nonexistent key" error when patching the Deployment seem to happen
because despite correctly unidling an app, sometimes Service update is
slow and after the refesh the unidler is still reached a second time.

As the unidler already removed this annotation when it re-`Patch()`
the `Deployment` it will return an error - sadly as we'd hope `Patch()`
would just return

> yep, you wanted me to remove that key, it's not there so it's all good"

We log this but ignore this error as it doesn't have an impact on the user.