Improved wording of "waiting for deployement" step
(agreed with Calum)

The message shown while we're waiting for a Deployment to be ready is
the message the user will see most as this is where the bulk of the
unidling time is spent.

The message is now showing 3 things:
1. Something good happened before: Replica restored
2. What's happening now: Starting app
3. Set the expectation for the user: This could take a few minutes...

Why waiting for a deploment can take long
=========================================

Because of the nature of k8s, this can take few seconds but it can also
take minutes if you're particularly "lucky", e.g.
- cluser is at capacity or almost at capacity
- cluster have to scale up and create a new node
- node needs to start
- your pod is scheduled to this new node
- the Docker image for the pod is new to the node and needs to be pulled
  (our tools images are particularly big)
- pod needs to start and this can take time as well