Fixed `undefined` message when connection to server falter
I finally got to the bottom of why sometimes the unidler client was
showing the weird `undefined` error message.

`onerror()` is invoked when there was an error sent by the server BUT ALSO
when the connection to the server is interrupted for whatever reason
(e.g. networking problems, we upgrade the unidler, whatever).

We need to handle connection errors (when `e.data` is `undefined`)
differently:

1. The error will not have a message (`e.data`) so there is no point in
   trying to display this `undefined` message to the user
2. We should NOT CLOSE the `EventSource` otherwise when the connection
   is re-enstablishe/retry works user will not see the new messages.

I'm confident this fixes [Issue #56: Unidler giving misleading messages].

[Issue #56: Unidler giving misleading messages]: https://github.com/ministryofjustice/analytics-platform/issues/95