Better handling of 0 replicas in annotation
Some few Deployments have the `mojanalytics.xyz/idled-at` annotation
with a odd number of original replicas.

This is probably caused by the idler when the replicas was set manually
to `0` by on of us.

Obviously it doesn't make any sense to set the replicas to `0` when
unidling an app, so now I log this oddity and set the replicas to `1`
in this cases.