Ignore "nonexistent key" Service Patch errors
Like we do in the case of the Deployment Patch. These errors are caused
when the unidler somehow tries to unidle an app more than once (e.g.
because the user refreshed or because the Ingress changes was slow
to take effect).

These errors are logged but users don't need to know about these or be
- in fact they usually only happen just after the app was unidled and
went away with a restart so it's better to handle them internally.

Thanks Callum for the report, very useful!