Using `replicas-when-unidled` annotation
The unidler used the `mojanalytics.xyz/idled-at` annotation which also
contains the time when the app was idled (comma-separated).
This was confusing and also fiddly as we had to split this string by `,`.

This change assumes that the `Deployment` has a `replicas-when-unidled`
annotation which only contains the number of replicas the `Deployment`
had when it was idled.

The idler currently doesn't currently write this annotation but it doesn't
really matter as:

1. the code assumes 1 replicas when the annotation is not present
2. our tools only have 1 replica anyway

the code also checks few things:

1. that the replicas read is at least 1, this because unidling and
   setting the replicas to something less than 1 doesn't make sense
2. it also checks if the `Deployment` replicas is already > 1, in that
   case doesn't patch it as it may be already unidled and/or in the
   process of being unidled and we can save the request to PATCH