Patch using strategic merge instead of JSONPatch
JSONPatch causes problems because of the way it's designed.
Its specification says:

> The target location MUST exist for the operation to be successful.

Unfortunately this means that a patch is not idempodent as we thought
and if it's applied 2+ times or if for whatever reason the resource
being patched is in an incosistent state it would error out.

Strategic merge seems to be much more robust as the patch is what you
want. This means we declare what the resource final state we want and
kubernetes will patch it accordingly - regardless of its state.

Part of ticket: https://trello.com/c/vr4LPcde/241-investigate-fix-idling-unidling-problems-caused-by-k8s-cluster-upgrade-new-version-of-k8s
JSONPatch RFC: https://tools.ietf.org/html/rfc6902