Use part before first "." as label selector
This is because k8s labels can't be longer than 63 chars and also they
can't end with a `.`

Introduced `UNIDLE_KEY_LABEL` environment variable to help with maintaining
compatibility with old `alpha` cluster.
This will default to `"host"` so that unidler will work as before in `alpha`.

Set `UNIDLE_KEY_LABEL=unidle-key` in new `prod` cluster to make the unidler
only use the part of the host before the first dot (".")

**NOTE**
This can be simplifie once we sunset `alpha` and we don't need to honour
the `host` label and old behaviour anymore.